made WebGlHost abstract & made classes for the different demos
