+++ docs/script_lighting.js
@@ -1,6 +1,6 @@
 import { CodeParcel } from "./code_parcel.js";
 import { GetBunnyModel } from "./models/stanfordbunny.js";
-import { WebGlHost } from "./webglhost.js";
+import { WebGlLightingDemo } from "./demos/lightingdemo.js";
 export class LightingDemo {
     constructor(model, parcel) {
         this.initializeDemo(model, parcel.vertexShaderCode, parcel.fragmentShaderCode, "lighting");
@@ -21,9 +21,7 @@ export class LightingDemo {
                 gl.enable(gl.DEPTH_TEST);
                 gl.clear(gl.COLOR_BUFFER_BIT);
                 gl.viewport(0, 0, canvas.width, canvas.height);
-                this.host = new WebGlHost(gl, model.vertices, model.indices, vertexShaderCode, fragmentShaderCode, pageTitle);
-                this.host.bindPositionAndNormal();
-                this.host.startRotationLoop();
+                this.host = new WebGlLightingDemo(gl, model.vertices, model.indices, vertexShaderCode, fragmentShaderCode, pageTitle);
             }
         }
     }
