+++ docs/webglhost.js
@@ -10,7 +10,9 @@ export class WebGlHost {
         this.pageString = pageString;
         this.initialiseWebGL();
         this.addButtonListener();
+        this.onloadHook();
     }
+    onloadHook() { }
     initialiseWebGL() {
         this.loadBuffers();
         this.loadShaders();
@@ -21,30 +23,10 @@ export class WebGlHost {
             this.clickEvent();
         });
     }
-    clickEvent() {
-        switch (this.pageString) {
-            case "loading":
-                this.reloadBuffers();
-                return;
-            case "camera":
-                this.reloadVertexShader();
-                return;
-            case "lighting":
-                this.reloadPixelShader();
-                return;
-        }
-    }
     renderCycle() {
         var gl = this.gl;
         gl.clear(gl.COLOR_BUFFER_BIT | gl.DEPTH_BUFFER_BIT);
         gl.drawElements(gl.TRIANGLES, this.indices.length, gl.UNSIGNED_SHORT, 0);
-    }
-    toNumArray(strings) {
-        var ans = new Array();
-        strings.forEach(function (s) {
-            ans.push(+s);
-        });
-        return ans;
     }
     loadBuffers() {
         var gl = this.gl;
@@ -91,79 +73,6 @@ export class WebGlHost {
             return shader;
         }
     }
-    reloadBuffers() {
-        // specific to loading page
-        var textAreaContent = this.getCodeSnippet();
-        this.vertices = this.toNumArray(textAreaContent.split(","));
-        this.loadBuffers();
-        this.loadingPageBindShaders();
-    }
-    reloadPixelShader() {
-        // specific to lighting page 
-        this.fragmentShaderCode = this.getCodeSnippet();
-        this.loadShaders();
-        this.bindPositionAndNormal();
-        this.renderCycle();
-    }
-    reloadVertexShader() {
-        // specific to camera page
-        this.vertexShaderCode = this.getCodeSnippet();
-        this.loadShaders();
-        this.updateSimpleCameraPosition();
-        this.renderCycle();
-    }
-    updateSimpleCameraPosition() {
-        if (!this.shaderProgram) {
-            return;
-        }
-        var gl = this.gl;
-        var uCamPosLocation = gl.getUniformLocation(this.shaderProgram, "u_cameraPosition");
-        gl.uniform3f(uCamPosLocation, this.cameraPosition[0], this.cameraPosition[1], this.cameraPosition[2]);
-    }
-    updateRotation() {
-        if (this.shaderProgram) {
-            var gl = this.gl;
-            var uRotationLocation = gl.getUniformLocation(this.shaderProgram, "u_rotation");
-            gl.uniform1f(uRotationLocation, this.rotation);
-            this.rotation = (this.rotation + 0.01) % 6.28;
-        }
-        this.renderCycle();
-        window.requestAnimationFrame(() => { this.updateRotation(); });
-    }
-    bindAttribute(attributeName, offset) {
-        if (!this.shaderProgram) {
-            return;
-        }
-        var gl = this.gl;
-        var coord = gl.getAttribLocation(this.shaderProgram, attributeName);
-        gl.vertexAttribPointer(coord, 3, gl.FLOAT, false, 6 * 4, offset);
-        gl.enableVertexAttribArray(coord);
-    }
-    //#endregion
-    updateCameraPositionOnKeyUp(event) {
-        if (this.moveCamera(event.key)) {
-            this.updateSimpleCameraPosition();
-        }
-        this.renderCycle();
-    }
-    moveCamera(key) {
-        if (key == 'd') {
-            this.cameraPosition[0] = this.cameraPosition[0] + 0.05;
-        }
-        else if (key == 'a') {
-            this.cameraPosition[0] = this.cameraPosition[0] - 0.05;
-        }
-        else if (key == 'w') {
-            this.cameraPosition[1] = this.cameraPosition[1] + 0.05;
-        }
-        else if (key == 's') {
-            this.cameraPosition[1] = this.cameraPosition[1] - 0.05;
-        }
-        else {
-            return false;
-        }
-        return true;
-    }
     getCodeSnippet() {
         const codeSection = document.getElementById("code");
         if (codeSection && codeSection instanceof HTMLTextAreaElement) {
@@ -181,12 +90,13 @@ export class WebGlHost {
         this.bindAttribute("a_normal", 3 * 4);
         this.renderCycle();
     }
-    setupCameraMovement() {
-        document.addEventListener('keyup', (event) => {
-            this.updateCameraPositionOnKeyUp(event);
-        }, false);
-    }
-    startRotationLoop() {
-        window.requestAnimationFrame(() => { this.updateRotation(); });
+    bindAttribute(attributeName, offset) {
+        if (!this.shaderProgram) {
+            return;
+        }
+        var gl = this.gl;
+        var coord = gl.getAttribLocation(this.shaderProgram, attributeName);
+        gl.vertexAttribPointer(coord, 3, gl.FLOAT, false, 6 * 4, offset);
+        gl.enableVertexAttribArray(coord);
     }
 }
